Data Control setting is seperated from Model Setting
