+++ src/components/Preferences/PreferencesModal.tsx
@@ -24,10 +24,12 @@ import ModelsSettings from "./ModelsSettings";
 import WebHandlersConfig from "./WebHandlersConfig";
 import DefaultSystemPrompt from "./DefaultSystemPrompt";
 import CustomizationSettings from "./CustomizationSettings";
+import DataControls from "./DataControls";
 import { MdOutlineDashboardCustomize, MdSignalCellularAlt } from "react-icons/md";
 import { IconType } from "react-icons";
 import { TbPrompt } from "react-icons/tb";
 import { FaRobot } from "react-icons/fa";
+import { FaDatabase } from "react-icons/fa";
 
 type PreferencesModalProps = {
   isOpen: boolean;
@@ -53,6 +55,7 @@ function PreferencesModal({ isOpen, onClose, finalFocusRef }: PreferencesModalPr
     { name: "System Prompt", icon: TbPrompt },
     { name: "Web Handlers", icon: MdSignalCellularAlt },
     { name: "Customization", icon: MdOutlineDashboardCustomize },
+    { name: "Data Controls", icon: FaDatabase },
   ];
 
   const handleSettingClick = (setting: Setting) => {
@@ -187,6 +190,7 @@ function PreferencesModal({ isOpen, onClose, finalFocusRef }: PreferencesModalPr
               {selectedSetting.name === "System Prompt" && <DefaultSystemPrompt />}
               {selectedSetting.name === "Web Handlers" && <WebHandlersConfig />}
               {selectedSetting.name === "Customization" && <CustomizationSettings />}
+              {selectedSetting.name === "Data Controls" && <DataControls />}
             </Box>
           </Flex>
         </ModalBody>
